requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-allauth==0.57.0
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3